+++ include/rtx.h
@@ -11,9 +11,9 @@
 int is_ray_in_bounds(ray3 ray, bounds3 bounds);
 int is_ray_in_bounds_seg(ray3 ray, bounds3 bounds);
 
-rtx_triangle* rtx_triangle_create(vec3 *triangle);
-rtx_triangle* rtx_triangle_create_param(vec3 *triangle, int alloc_lightmap);
-rtx_triangle* rtx_triangle_create_discrete(vec3 a, vec3 b, vec3 c);
+rtx_triangle* rtx_triangle_create(dvec3 *triangle);
+rtx_triangle* rtx_triangle_create_param(dvec3 *triangle, int alloc_lightmap);
+rtx_triangle* rtx_triangle_create_discrete(dvec3 a, dvec3 b, dvec3 c);
 void rtx_triangle_destroy(rtx_triangle *triangle);
 
 void vec_rtx_triangle_add(vec_rtx_triangle *vec, rtx_triangle *to_add);
